Moved opponent services into groovybones.opponent/ package
